+++ src/fiflo.c
@@ -8,7 +8,7 @@
 #include "input.h"
 #include "window.h"
 
-void options(const char* arg)
+bool options(const char* arg)
 {
 	if(!strcmp(arg, "-h") || !strcmp(arg, "--help"))
 	{
@@ -17,34 +17,48 @@ void options(const char* arg)
 
 		"Options:      Description:",
 		"-h, --help    Show a program help.",
-		"-v, --version Display information about the version You use.");
-		exit(0);
+		"-v, --version Display information about this version.");
+		return false;
 	}
 	else if(!strcmp(arg, "-v") || !strcmp(arg, "--version"))
 	{
 		printf("%s\n%s\n%s\n%s\n%s\n",
 		"|``",
 		"|``",
-		"fiflo v2.3.0 (WIP)",
+		"fiflo v3.0.0 (WIP)",
 		"(C) 2018-2019 mattmaniak, MIT License",
 		"https://gitlab.com/mattmaniak/fiflo.git");
-		exit(0);
+		return false;
 	}
+	return true;
 }
 
-noreturn void run(const char* arg)
+bool run(const char* arg)
 {
 	char   pressed = '\0'; // Initializer.
 	Buff_t Buff;
 
-	buffer.init(&Buff);
-	file.set_name(&Buff, arg);
-	file.load(&Buff);
+	if(!buffer.init(&Buff))
+	{
+		return false;
+	}
+	if(!file.set_name(&Buff, arg))
+	{
+		return false;
+	}
+	if(!file.load(&Buff))
+	{
+		return false;
+	}
 
 	for(;;) // The main program loop.
 	{
 		file.get_git_branch(&Buff);
-		input.parse_key(&Buff, pressed);
+		if(input.parse_key(&Buff, pressed))
+		{
+			// buffer.free_all(&Buff);
+			// return false;
+		}
 		window.render(&Buff);
 
 		pressed = input.getch(&Buff);
@@ -57,22 +71,25 @@ int main(const int argc, const char** argv)
 	if(argc > 2)
 	{
 		fputs("Only one additional arg can be passed.\n", stderr);
-		return 1;
+		return -1;
 	}
 
 	if(argv[1] == NULL)	// Sets the default basename and starts.
 	{
-		run("");
+		return run("");
 	}
 	else if(strlen(argv[1]) <= PATH_MAX)
 	{
-		options(argv[1]);
-		run(argv[1]);
+		if(!options(argv[1]))
+		{
+			return 0;
+		}
+		return run(argv[1]);
 	}
 	else
 	{
 		fprintf(stderr, "Maximum length of a parameter is %u\n", PATH_MAX);
-		return 1;
+		return -1;
 	}
 }