+++ src/fiflo.h
@@ -7,10 +7,10 @@
 #include <stdlib.h>
 
 // Program parameters, eg. "--help".
-void options(const char* arg);
+bool options(const char* arg);
 
 // Some initializers and the main program loop.
-void run(const char* arg);
+bool run(const char* arg);
 
 int main(const int argc, const char** argv);