+++ src/file.h
@@ -15,13 +15,13 @@ types. For better readability they are named as their definitions. */
 typedef const struct
 {
 	// Sets the filename. If basename is passed as arg actent path will be added.
-	Buff_t* (*set_name)(Buff_t*, const char*);
+	bool (*set_name)(Buff_t*, const char*);
 
 	// When the CTRL^O is pressed, there is possibility to basically edit the fname.
 	Buff_t* (*live_edit_name)(Buff_t*, char);
 
 	// Reads the file using "recognize_char".
-	Buff_t* (*load)(Buff_t*);
+	bool (*load)(Buff_t*);
 
 	// Puts the whole text to the file.
 	Buff_t* (*save)(Buff_t*);