+++ doc/CHANGELOG.md
@@ -7,9 +7,10 @@ Changelog contains only new tangible features. To see what's new in source
 files use git-show or git-log.
 
 ## Versions:
-### v2.3.0 (WIP):
+### v3.0.0 (WIP):
 - 80th column indicator,
-- display git branch name.
+- display git branch name,
+- return-based API.
 
 ### v2.2.2 (2018, December 22):
 - Repaired positions of line numbers when the line has same width as the